requirements.txt
anyio==3.7.1
APScheduler==3.6.3
blinker==1.6.2
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Flask==2.3.2
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-speech==2.21.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
Werkzeug==2.3.6
